+++ packages/patrol_log/lib/src/emojis.dart
@@ -3,6 +3,7 @@ class Emojis {
   static const testStart = '🧪';
   static const success = '✅';
   static const failure = '❌';
+  static const flaky = '⚠️';
   static const skip = '⏩';
   static const log = '📝';
   static const report = '📊';